+++ src/main/kotlin/com/github/sleepypanda/feesh/events/models/BaitConsumedEvent.kt
@@ -1,0 +1,7 @@
+package com.github.sleepypanda.feesh.events.models
+
+data class BaitConsumedEvent(
+    val baitName: String,
+    val baitDisplayName: String,
+    val baitId: String
+)